Created first calc function
In this code version, I have removed all the classes for the functions I created. I removed the class because they were getting confusing, and I believe that the way I tried to use classes was unnecessary. In this version of the code, I also completed writing the first function for the area of a circle, including each display mode the user may want (equation or default).