added volume of cylinder calc
In this update, I added the volume of a cylinder calculation. In this function, the radius and heights float/int of the cylinder are given by the user, and the volume is calculated. The program then rounds the value and prints it in either equation view or default view (depending on the mode selected).