added area of triangle calc
in this update I have added the function triangle_area.  In this function, the base and height of the base float/int of the triangle are given by the user, and the area is calculated. The program then rounds the value and prints it in either equation view or default view (depending on the mode selected). In this update I also made a few small changes and realized my equation view wasn't working properly and fixed it.