added volume of sphere calc
In this update, I added the volume of a sphere calculation. In this function, the user gives the radius float/int of the sphere, and the volume is calculated. The program then rounds the value and prints it in either equation view or default view (depending on the mode selected).