update for pytest
In this commit, I realized that the way I made my program previously wouldn't really work with pytest (because I couldn't pass values to the functions). So, I have changed all the functions to instead get the value (input) from the if statements of the main part and pass it to the functions. This should allow me to properly pass values to the functions from the pytest file for testing.