Final update to pytest
In this commit, I have added the last three functions used to test the functions in the main program (basically the same as the first two). I have also added the rest of the comments to the code and cleaned it up a bit. I checked to ensure that all the calculations were working properly with pytest, and I got no errors (5 passed in 0.02 sec).