final changes
In this update, I realized that I still didn't have my menu (level 0 and 1) set up separately, so I moved level 0 outside of the loop to only be displayed at the beginning of the program. I have also added the rest of my comments and cited material. The last thing I changed in this commit was I added units (^2 or ^3) to the calculations (I didn't add m^2 or cm^2 on the result because I wanted to let the user choose what units they want to use)